stripe and paypal components moved into their own files respectively, as per Dom's suggestion
