+++ app/services/ContributionOphanService.scala
@@ -5,11 +5,10 @@ import actions.CommonActions.ABTestRequest
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
 import cats.data.EitherT
+import cats.syntax.EitherSyntax
 import cats.syntax.either._
 import com.gu.acquisition.services.{OphanService, OphanServiceError}
-import com.gu.stripe.Stripe.Charge
 import com.paypal.api.payments.Payment
-import controllers.forms.ContributionRequest
 import controllers.httpmodels.CaptureRequest
 import ophan.thrift.componentEvent.ComponentType
 import ophan.thrift.event.{AbTest, AbTestInfo, Acquisition, AcquisitionSource}
@@ -75,7 +74,7 @@ object ContributionOphanService {
   /**
     * Mixin to facilitate creating acquisition submission builders.
     */
-  trait AcquisitionSubmissionBuilderUtils {
+  trait AcquisitionSubmissionBuilderUtils extends EitherSyntax {
 
     def tryField[A](name: String)(a: => A): Either[OphanServiceError, A] =
       Either.catchNonFatal(a).leftMap { _ =>
@@ -86,149 +85,6 @@ object ContributionOphanService {
       import com.gu.acquisition.syntax._
       val abTestInfo = native.asAbTestInfo
       nonNative.map(abTest => AbTestInfo(abTestInfo.tests + abTest)).getOrElse(abTestInfo)
-    }
-  }
-}
-
-case class StripeAcquisitionComponents(charge: Charge, request: ABTestRequest[ContributionRequest])
-
-object StripeAcquisitionComponents {
-
-  implicit object stripeAcquisitionSubmissionBuilder
-    extends AcquisitionSubmissionBuilder[StripeAcquisitionComponents]
-    with AcquisitionSubmissionBuilderUtils {
-
-    override def buildAcquisition(components: StripeAcquisitionComponents): Either[OphanServiceError, Acquisition] = {
-      import components._
-
-      Either.right(
-        Acquisition(
-          product = ophan.thrift.event.Product.Contribution,
-          paymentFrequency = ophan.thrift.event.PaymentFrequency.OneOff,
-          currency = charge.currency,
-          amount = charge.amount,
-          amountInGBP = None, // Calculated at the sinks of the Ophan stream
-          paymentProvider = Option(ophan.thrift.event.PaymentProvider.Stripe),
-          campaignCode = Some(Set(request.body.intcmp, request.body.cmp).flatten),
-          abTests = Some(abTestInfo(request.testAllocations, request.body.abTest)),
-          countryCode = Some(charge.source.country),
-          referrerPageViewId = request.body.refererPageviewId,
-          referrerUrl = request.body.refererUrl,
-          componentId = request.body.componentId,
-          componentTypeV2 = request.body.componentType,
-          source = request.body.source
-        )
-      )
-    }
-
-    def buildOphanIds(components: StripeAcquisitionComponents): Either[OphanServiceError, OphanIds] = {
-      import components._
-      tryField("ophanBrowserId")(request.body.ophanBrowserId.get)
-        .map { browserId =>
-          OphanIds(browserId, request.body.ophanPageviewId, request.body.ophanVisitId)
-        }
-    }
-  }
-}
-
-object PaypalAcquisitionComponents {
-
-  case class Execute(payment: Payment, request: Execute.RequestData)
-
-  object Execute {
-
-    case class RequestData(
-        cmp: Option[String],
-        intCmp: Option[String],
-        refererPageviewId: Option[String],
-        refererUrl: Option[String],
-        ophanPageviewId: Option[String],
-        ophanBrowserId: Option[String],
-        ophanVisitId: Option[String],
-        componentId: Option[String],
-        componentType: Option[ComponentType],
-        source: Option[AcquisitionSource],
-        abTest: Option[AbTest],
-        testAllocations: Set[Allocation]
-    )
-
-    implicit object paypalAcquisitionSubmissionBuilder
-      extends AcquisitionSubmissionBuilder[Execute] with AcquisitionSubmissionBuilderUtils {
-
-      def buildOphanIds(components: Execute): Either[OphanServiceError, OphanIds] = {
-        import components._
-        for {
-          browserId <- tryField("ophanBrowserId")(request.ophanBrowserId.get)
-          pageviewId <- tryField("ophanPageviewId")(request.ophanPageviewId.get)
-        } yield OphanIds(browserId, pageviewId, request.ophanVisitId)
-      }
-
-      def buildAcquisition(components: Execute): Either[OphanServiceError, Acquisition] = {
-        import com.gu.acquisition.syntax._
-        import components._
-        for {
-          amount <- tryField("amount")(payment.getPaymentInstruction.getAmount.getValue.toDouble)
-        } yield {
-          Acquisition(
-            product = ophan.thrift.event.Product.Contribution,
-            paymentFrequency = ophan.thrift.event.PaymentFrequency.OneOff,
-            currency = payment.getPaymentInstruction.getAmount.getCurrency,
-            amount = amount,
-            amountInGBP = None, // Calculated at the sinks of the Ophan stream
-            paymentProvider = Some(ophan.thrift.event.PaymentProvider.Paypal),
-            campaignCode = Some(Set(request.intCmp, request.cmp).flatten),
-            abTests = Some(request.testAllocations.asAbTestInfo),
-            countryCode = Some(payment.getPayer.getPayerInfo.getCountryCode),
-            referrerPageViewId = request.refererPageviewId,
-            referrerUrl = request.refererUrl,
-            componentId = request.componentId,
-            componentTypeV2 = request.componentType,
-            source = request.source
-          )
-        }
-      }
-    }
-  }
-
-  case class Capture(payment: Payment, request: ABTestRequest[CaptureRequest])
-
-  object Capture {
-
-    implicit object paypalAcquisitionSubmissionBuilder
-      extends AcquisitionSubmissionBuilder[Capture] with AcquisitionSubmissionBuilderUtils {
-
-      override def buildOphanIds(components: Capture): Either[OphanServiceError, OphanIds] = {
-        import components._
-        for {
-          browserId <- tryField("ophanBrowserId")(request.body.ophanBrowserId.get)
-          pageviewId <- tryField("ophanPageviewId")(request.body.ophanPageviewId.get)
-        } yield OphanIds(browserId, pageviewId, visitId = None)
-      }
-
-      override def buildAcquisition(components: Capture): Either[OphanServiceError, Acquisition] = {
-        import components._
-
-        for {
-          amount <- tryField("amount")(payment.getPaymentInstruction.getAmount.getValue.toDouble)
-        } yield {
-          Acquisition(
-            product = ophan.thrift.event.Product.Contribution,
-            paymentFrequency = ophan.thrift.event.PaymentFrequency.OneOff,
-            currency = payment.getPaymentInstruction.getAmount.getCurrency,
-            amount = amount,
-            amountInGBP = None, // Calculated at the sinks of the Ophan stream
-            paymentProvider = Some(ophan.thrift.event.PaymentProvider.Paypal),
-            campaignCode = Some(Set(request.body.cmp, request.body.intCmp).flatten),
-            abTests = Some(abTestInfo(request.testAllocations, request.body.abTest)),
-            countryCode = Some(payment.getPayer.getPayerInfo.getCountryCode),
-            referrerPageViewId = request.body.refererPageviewId,
-            referrerUrl = request.body.refererUrl,
-            componentId = request.body.componentId,
-            componentTypeV2 = request.body.componentType,
-            source = request.body.source
-          )
-        }
-      }
     }
   }
 }
